code/python/week-3/pyproject.toml
[project]
name = "week-3-rag-fundamentals"
version = "0.1.0"
description = "RAG fundamentals: Gemini File Search and DIY with Postgres"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "docling>=2.0.0",
    "docling-core[chunking]>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pgvector>=0.2.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.0.0",
]
